requirements.txt
fastapi
uvicorn
pymongo
langchain
langchain-openai
langchain-google-genai
python-dotenv
pydantic
requests
websockets